[Backport 2.6][#8360] ybase: Correctly compare running workers
Summary:
When investigating QLTransactionTest.IntentsCleanupAfterRestart failure, I noticed the following additional log:
```
I0509 16:29:40.284875 399244 priority_thread_pool.cc:584] We already have 10 - 0 - 7 >= 18446744073709551615 workers running, we could not run a new worker.
```
workers_.size() : 10
paused_workers_ : 0
free_workers_.size() : 7
max_running_tasks_ : 18446744073709551615

Note 18446744073709551615 is actually 2^64 − 1
The wraparound led to wrong result.

This revision changes the related fields in the comparison to int64_t so that there is no such wraparound.

Also in case compactions are disabled in a test, previously we were incorrectly setting the priority thread pool size to an infinite number by passing -1 as a size_t parameter. Now we are setting it to 1 in that case.

Test Plan:
Jenkins: rebase: 2.6
yb_build.sh --cxx-test client_ql-transaction-test --gtest_filter QLTransactionTest.IntentsCleanupAfterRestart
yb_build.sh --cxx-test client_ql-dml-test --gtest_filter QLDmlTest.RangeFilter

Reviewers: mbautin

Reviewed By: mbautin

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D11544

+++ src/yb/util/priority_thread_pool.h
@@ -57,7 +57,7 @@ class PriorityThreadPoolTask {
 // Tasks submitted to this pool have assigned priority and are picked from queue using it.
 class PriorityThreadPool {
  public:
-  explicit PriorityThreadPool(size_t max_running_tasks);
+  explicit PriorityThreadPool(int64_t max_running_tasks);
   ~PriorityThreadPool();
 
   // Submit task to the pool.
